requirements.txt
aiogram==3.13.0
python-dotenv
pytest-asyncio
